8.14 Bugfix

+++ src/controllers/videoController.js
@@ -38,10 +38,16 @@ export const watch = async (req, res) => {
 export const getEdit = async (req, res) => {
   const { id } = req.params;
   const video = await Video.findById(id);
+  const {
+    user: { _id },
+  } = req.session;
   if (!video) {
     return res
       .status(404)
       .render("404.pug", { pageTitle: "Video not found 😕" });
+  }
+  if (String(video.owner) !== _id) {
+    return res.status(403).redirect("/");
   }
   return res.render("videos/edit.pug", {
     pageTitle: `Edit ${video.title}`,
@@ -51,24 +57,44 @@ export const getEdit = async (req, res) => {
 export const postEdit = async (req, res) => {
   const { id } = req.params;
   const { title, description, hashtags } = req.body;
-  const video = await Video.exists({ _id: id });
+  const {
+    user: { _id },
+  } = req.session;
+  const video = await Video.findById(id);
   if (!video) {
     return res
       .status(404)
       .render("404.pug", { pageTitle: "Video not found 😕" });
+  }
+  if (String(video.owner) !== _id) {
+    return res.status(403).redirect("/");
   }
   await Video.findByIdAndUpdate(id, {
     title,
     description,
     hashtags: Video.formatHashtags(hashtags),
   });
-
   return res.redirect(`/videos/${id}`);
 };
 
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
+  const video = await Video.findById(id);
+  const {
+    user: { _id },
+  } = req.session;
+  const user = await User.findById(_id);
+  if (!video) {
+    return res
+      .status(404)
+      .render("404.pug", { pageTitle: "Video not found 😕" });
+  }
+  if (String(video.owner) !== _id) {
+    return res.status(403).redirect("/");
+  }
   await Video.findByIdAndDelete(id);
+  user.videos.splice(user.videos.indexOf(id), 1);
+  user.save();
   return res.redirect("/");
 };
